broke out the hurricane tracks into a separate component
